requirements.txt
langchain
langchain-core
langchain-community
langchain-text-splitters
langchain-huggingface

langchain-openai
langchain-mistralai
openai
mistralai

langchain-pinecone
pinecone

sentence-transformers
pypdf

python-dotenv

fastapi
uvicorn

pydantic
httpx

python-multipart

tqdm
requests